Address PR review / Search inside box

+++ src/core/search/ast_expr.cc
@@ -20,7 +20,7 @@ AstRangeNode::AstRangeNode(double lo, bool lo_excl, double hi, bool hi_excl)
     : lo{lo_excl ? nextafter(lo, hi) : lo}, hi{hi_excl ? nextafter(hi, lo) : hi} {
 }
 
-AstGeoNode::AstGeoNode(double lat, double lon, size_t radius, std::string unit)
+AstGeoNode::AstGeoNode(double lat, double lon, double radius, std::string unit)
     : lat(lat), lon(lon), radius(radius), unit(std::move(unit)) {
 }
